requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2.2
django-jazzmin==2.6.0
django-json-widget==1.1.1
django-localflavor==4.0
django-multiupload==0.6.1
django-otp==1.2.1
django-shortuuidfield==0.1.3
future==0.18.3
idna==3.4
jsoneditor==1.6.0
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pyperclip==1.8.2
python-stdnum==1.18
PyYAML==6.0
razorpay==1.3.0
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.3
